requirements.txt
asgiref==3.11.0
colorama==0.4.6
Django==5.2.9
djangorestframework==3.16.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
sqlparse==0.5.5
tzdata==2025.3
asgiref==3.11.0
colorama==0.4.6
Django==5.2.9
djangorestframework==3.16.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
sqlparse==0.5.5
tzdata==2025.3
asgiref==3.11.0
attrs==25.4.0
colorama==0.4.6
Django==5.2.9
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
